Foundation created, type and inputs for answers
I created the basic foundation for the program. A simple updating title with what type you're guessing for.

Then the program creates inputs based on how many weaknesses the current type has.

The user can enter by typing, when a correct answer is given the input will disable and turn green, if answer is incorrect it will clear the input so the user can retry.

Upon filling in all answers correctly the program will reset and rerandomize.