Grid gametype full functionality + code optimization
Added full functionality for the grid/select gametype. Added a counter to grid so it shows how many the user still needs to select.

Changed some code around so the functions for grid and input types are the same.

Changed files:
Removed all the input/grid functionality to it's own script. Now gamefunctions and play functions are seperated for readability.